==> how_to_download_sp500_data/1 how to get sp500 tickers.ipynb <==
# import requests

import requests
url = 'http://en.wikipedia.org/wiki/List_of_S%26P_500_companies'
#url = 'http://en.wikipedia.org/wiki/List_of_S%26P_500_company'
resp = requests.get(url)
#print("응답 코드 :", resp.status_code) # 200 이면 정상

import requests
url = 'http://en.wikipedia.org/wiki/List_of_S%26P_500_companies'
#url = 'http://en.wikipedia.org/wiki/List_of_S%26P_500_company'
resp = requests.get(url)
if resp.status_code == requests.codes.ok:
    print('정상입니다')
else:
    print('문제가 생겼습니다. [에러코드 {}]'.format(resp.status_code))

import requests
url = 'http://en.wikipedia.org/wiki/List_of_S%26P_500_companies'
#url = 'http://en.wikipedia.org/wiki/List_of_S%26P_500_company'
resp = requests.get(url)
resp.raise_for_status()
print('SP500 티커들을 모아봅니다')

type(resp) # requests.models.Response

resp

type(resp.text)

resp.text[:100] # the text-based content of web files
#resp.text # the text-based content of web files

# import bs4 as bs

import bs4 as bs
soup = bs.BeautifulSoup(resp.text, 'lxml')

type(soup) # bs4.BeautifulSoup

# find vs findAll - find

[HTML Tags Ordered Alphabetically](https://www.w3schools.com/TAGS/default.ASP)

table = soup.find('table', {'class': 'wikitable sortable'})

type(table) # bs4.element.Tag

# find vs findAll - findAll

It doesn't download data for stocks "BRK.B", and "BR.B".

Your problem is Yahoo Finance doesn't use the "." notation to track shares of different classes. So, "BRK.B" and "BR.B" are actually "BRKB" and "BRB". 

My Answer:
Your problem is Yahoo Finance doesn't use the "." notation to track shares of different classes. So, "BRK.B" and "BR.B" are actually "BRK-B" and "BR-B", not "BRKB" and "BRB". 

[stackoverflow](https://stackoverflow.com/questions/45379155/downloading-data-from-yahoo-finance-with-a-dash-or-dot-in-the-ticker)

type(table.findAll('tr'))

type(table.findAll('tr')[:])

type(table.findAll('tr')[0])

type(table.findAll('tr')[0].findAll('td'))

type(table.findAll('tr')[0].findAll('td')[:])

try:
    type(table.findAll('tr')[0].findAll('td')[0])
except IndexError as err:
    print(err)

tickers = []
for row in table.findAll('tr')[1:]:
    ticker = row.find('td').text[:-1]
    ticker = ticker.replace('.', '-')
    tickers.append(ticker)
print(tickers)

import pandas as pd
tickers = pd.DataFrame(tickers)
print(tickers)

import os
try:
    os.makedirs('../data/sp500')
except:
    pass

tickers.to_csv('../data/sp500/tickers.csv')

tickers = pd.read_csv('../data/sp500/tickers.csv', usecols=[1])
tickers

# Reference

[List of S&P 500 companies](http://en.wikipedia.org/wiki/List_of_S%26P_500_companies)

[How To Work with Web Data Using Requests and Beautiful Soup with Python 3](https://www.digitalocean.com/community/tutorials/how-to-work-with-web-data-using-requests-and-beautiful-soup-with-python-3)

[Getting all company pricing data in the S&P 500 - Python Programming for Finance p.6](https://www.youtube.com/watch?v=baCAFPHb1o4)

[파이썬 코딩 무료 강의 (활용편3) - 웹 크롤링? 웹 스크래핑! 제가 가진 모든 비법을 알려드리겠습니다. [나도코딩]](https://www.youtube.com/watch?v=yQ20jZwDjTE)

[[HTML Tags Ordered Alphabetically](https://www.w3schools.com/TAGS/default.ASP)](https://www.w3schools.com/TAGS/default.ASP)

[w3schools](https://www.w3schools.com/)
