==> how_to_download_sp500_data/2 how to download sp500 data.ipynb <==
import pandas as pd
tickers = pd.read_csv('../data/sp500/tickers.csv', usecols=[1])
tickers

tickers = list(tickers.iloc[:,0])
tickers

import yfinance as yf
#import time
import os
tickers_len = len(tickers)
for count, ticker in enumerate(tickers):
    if not os.path.exists('../data/sp500/{}.csv'.format(ticker)):
        stock = yf.Ticker(ticker)
        data = stock.history(period='max')
        data.to_csv('../data/sp500/{}.csv'.format(ticker))
        print('{}/{} : {} Downloaded'.format(count+1, tickers_len, ticker))
        #time.sleep(0.5)
    else:
        print('{}/{} : {} Downloaded'.format(count+1, tickers_len, ticker))

import yfinance as yf
import time
import pandas as pd
import os
tickers_len = len(tickers)
for count, ticker in enumerate(tickers):
    data = pd.read_csv('../data/sp500/{}.csv'.format(ticker))
    if data.empty:
        stock = yf.Ticker(ticker)
        data = stock.history(period='max')
        data.to_csv('../data/sp500/{}.csv'.format(ticker))
    print('{}/{} : {} Downloaded'.format(count+1, tickers_len, ticker))

# Reference

[How To Work with Web Data Using Requests and Beautiful Soup with Python 3](https://www.digitalocean.com/community/tutorials/how-to-work-with-web-data-using-requests-and-beautiful-soup-with-python-3)
