==> how_to_download_sp500_data/3 how to combine sp500 data.ipynb <==
import pandas as pd
tickers = pd.read_csv('../data/sp500/tickers.csv', usecols=[1])
tickers

tickers = list(tickers.iloc[:,0])
tickers

df_close = pd.DataFrame()

for count, ticker in enumerate(tickers):
    
    df = pd.read_csv('../data/sp500/{}.csv'.format(ticker), 
                     index_col='Date',
                     parse_dates=True)
    df.rename(columns={'Close':ticker}, inplace=True)
    
    # 0, or ‘index’ : Drop rows which contain missing values. (Default)
    # 1, or ‘columns’ : Drop columns which contain missing value.
    df.drop(['Open','High','Low','Volume','Dividends','Stock Splits'],
            axis=1,
            inplace=True)
    
    # 0, or ‘index’ : Drop rows which contain missing values. (Default)
    # 1, or ‘columns’ : Drop columns which contain missing value.
    df.dropna(axis=0, inplace=True) # 배당 혹은 주식분할로 생기는 추가적인 가로열 제거
    
    if df_close.empty:
        df_close = df
    else:
        df_close = df_close.join(df, how='outer')
        
    if count % 10 == 0:
        print(count)

df_close.to_csv('../data/sp500/Close.csv')
df_close = pd.read_csv('../data/sp500/Close.csv', index_col='Date', parse_dates=True)

df_close.head()

df_close.tail()

# Reference

[How To Work with Web Data Using Requests and Beautiful Soup with Python 3](https://www.digitalocean.com/community/tutorials/how-to-work-with-web-data-using-requests-and-beautiful-soup-with-python-3)
