==> how_to_download_sp500_data/4 how to compute daily returns.ipynb <==
import pandas as pd
df_close = pd.read_csv('../data/sp500/Close.csv', 
                       index_col='Date', 
                       parse_dates=True)
df_close.tail()

$$
\frac{S_t-S_{t-1}}{S_{t-1}}
$$

df_pct_change = df_close.pct_change()
df_pct_change.to_csv('../data/sp500/Pct_Change.csv')
df_pct_change = pd.read_csv('../data/sp500/Pct_Change.csv', index_col='Date', parse_dates=True)

df_pct_change.head()

df_pct_change.tail()

# Reference

[How to find Percentage Change in pandas](https://kanoki.org/2019/09/29/how-to-find-percentage-change-in-pandas/)

[pandas.DataFrame.pct_change](https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.pct_change.html)
