==> how_to_download_sp500_data/7 how to prepare data for train.ipynb <==
import numpy as np; np.random.seed(0)
import pandas as pd
from collections import deque

target_index_size = 10  # 포트폴리오 자산 갯수
n_days_train = 21*6    # 6개월 트래인 기간
n_days_test = 21       # 1개월 테스트 기간
batch_size = 1000      # 한번 트래인시 필요한 데이터 갯수

tickers = pd.read_csv('../data/sp500/tickers.csv', usecols=[1])
n_assets = tickers.shape[0]
print(f'총 자산 수 : {n_assets}')
print('티커 : {}'.format(list(tickers.iloc[:,0])))

target_index = list(np.random.choice(np.arange(n_assets), target_index_size, replace=False))
train_index = []
for i in range(n_assets-1):
    if i not in target_index:
        train_index.append(i)
        
tickers_target = list(tickers.iloc[target_index, 0]) 
tickers_train = list(tickers.iloc[train_index, 0])
        
print('타겟 자산 수 : {}'.format(len(target_index)))
print('타겟 자산 인덱스 : {}'.format(target_index))
print('타겟 티커 : {}'.format(tickers_target), end='\n\n')

print('트래인 자산 수 : {}'.format(len(train_index)))
print('트래인 자산 인덱스 : {}'.format(train_index))
print('트래인 티커 : {}'.format(tickers_train))

df_pct_change = pd.read_csv('../data/sp500/Pct_Change.csv', index_col='Date', parse_dates=True)
n_days = df_pct_change.shape[0] 
print(df_pct_change.shape)

def give_me_x_y():
    
    x = []
    y = []

    for i in range(batch_size):

        while True: 

    #         target_index_size = 20 # 포트폴리오 자산 갯수
    #         n_days_train = 21*6    # 6개월 트래인 기간
    #         n_days_test = 21       # 1개월 테스트 기간
    #         batch_size = 1000      # 한번 트래인시 필요한 데이터 갯수
            start_day_x = np.random.choice(np.arange(n_days-n_days_train-n_days_test), 1)[0] 
            end_day_x = start_day_x + n_days_train - 1 
            start_day_y = start_day_x + n_days_train
            end_day_y = start_day_x + n_days_train + n_days_test - 1 

            asset_index = list(np.random.choice(train_index, target_index_size, replace=False))

            data_x = df_pct_change.iloc[start_day_x:end_day_x+1,asset_index]
            data_y = df_pct_change.iloc[start_day_y:end_day_y+1,asset_index]

            if not data_x.isnull().any().any():
                if not data_y.isnull().any().any():
                    try:
                        x.append(data_x.values.reshape((-1,)))
                        y.append(data_y.values.reshape((-1,)))
                        break
                    except:
                        pass

    x = np.array(x)
    y = np.array(y)
    print(f'트래인 셋 인풋 사이즈 : {x.shape}')
    print(f'트래인 셋 아웃풋 사이즈 : {y.shape}')
    
    return x, y

x, y = give_me_x_y() 
print(f'첫번째 트래인 셋 인풋 : {x[0]}')
print(f'첫번째 트래인 셋 아웃풋 : {y[0]}')

for i in range(3):
    give_me_x_y() 

buffer = deque(maxlen=3)
for i in range(10):
    buffer.append(give_me_x_y())
    #print(buffer)
    print(len(list(buffer)))
